Fix: small-caps wrongly rendered in Linux
Small caps with Palatino Linotye in Linux are rendered too small,
especially for digits which appear in the same size as lower case
characters. This commit fixes at least the rendering for digits.

I am not sure if this is an issue of freetype, infinality or Palatino
Linotype, but it works in Windows.